+++ includes/Light/PointLight.h
@@ -6,17 +6,17 @@
 #include "Light.h"
 
 struct PointLight : Light {
-    glm::vec3 position, attenuation;
+    glm::dvec3 position, attenuation;
 
-    PointLight(const glm::vec3 &position) : position(position) {}
+    PointLight(const glm::dvec3 &position) : position(position) {}
 
-    [[nodiscard]] glm::vec3 get_intensity(const glm::vec3 &point) const override;
+    [[nodiscard]] glm::dvec3 get_intensity(const glm::dvec3 &point) const override;
 
-    [[nodiscard]] glm::vec3 get_on_light_direction(const glm::vec3 &point) const override;
+    [[nodiscard]] glm::dvec3 get_on_light_direction(const glm::dvec3 &point) const override;
 
-    [[nodiscard]] float get_distance(const glm::vec3 &point) const override;
+    [[nodiscard]] double get_distance(const glm::dvec3 &point) const override;
 
-    void set_attenuation(const glm::vec3 &);
+    void set_attenuation(const glm::dvec3 &);
 };
