+++ src/Object/Object.cpp
@@ -9,7 +9,7 @@ Ray Object::prepare_ray(const Ray &ray) const {
     return new_ray;
 }
 
-ObjectIntersection Object::get_intersection_info(const Ray &ray, glm::vec3 &normal, float t) const {
+ObjectIntersection Object::get_intersection_info(const Ray &ray, glm::dvec3 &normal, double t) const {
     bool is_inside = glm::dot(normal, ray.dir) > 0;
     if (is_inside) normal *= -1;
     normal = rotation * normal;
